fix so openmpi with verbs works outside build server
Outside of the build server the module would not load.
Spider could find it but it would not load.
I think this may be due to the previous openmpi modules
being built as Core and I was trying to build this as Comp.
I am now building this as Core and this requires removing
fortran openmpi support.