fix Unable to create immutable header region error
There is a limitation on the number of files possible in
an rpm. When this limit is reached the make step will fail
with the following error:
    Unable to create immutable header region.
This problem occurred without any changes. I think the
issue was caused by new files pulled in for the existing
packages installed. To work around this problem I removed
the beautifulsoup4 package.